requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.5
django-allauth==65.11.1
idna==3.10
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
stripe==12.5.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
